+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v1.3.0
+- [feat] add `go-cache`
+- [chore] use `owl` instead of `viper` dep
+
 ## v1.2.0
 - [feat]: dashboard add auth middleware
 - [refactor]: delete `cobra` cli lib